requirements.txt
Fabric
lxml
python-dateutil
dataset==0.3
nose==1.3.0
